+++ R/messages.R
@@ -41,16 +41,14 @@ info <- function(
   msg_types = getOption("msgr.types"),
   log_path  = getOption("msgr.log_path"))
 {
-  {
-    (is_natural(level) && level <= 10) ||
-      stop("'level' must be an integer between 1 and 10: ", level)
-    (is_natural(msg_level) && msg_level <= 10) ||
-      stop("'msg_level' must be an integer between 1 and 10: ", msg_level)
-    ((is_null(msg_types) || is_character(msg_types)) && is_in(msg_types, c("INFO", "WARNING", "ERROR"))) ||
-      stop("'msg_types' must be NULL (no messages) or a character vector containing \"INFO\", \"WARNING\" or \"ERROR\": ", msg_types)
-    is_string(log_path) ||
-      stop("'log_path' must be a string: ", log_path)
-  }
+  (is_scalar_integerish(level) && isTRUE(level > 0) && isTRUE(level <= 10)) ||
+    stop("'level' must be an integer between 1 and 10: ", level)
+  (is_scalar_integerish(msg_level) && isTRUE(msg_level > 0) && isTRUE(msg_level <= 10)) ||
+    stop("'msg_level' must be an integer between 1 and 10: ", msg_level)
+  ((is_null(msg_types) || is_character(msg_types)) && is_in(msg_types, c("INFO", "WARNING", "ERROR"))) ||
+    stop("'msg_types' must be NULL (no messages) or a character vector containing \"INFO\", \"WARNING\" or \"ERROR\": ", msg_types)
+  (is_scalar_character(log_path)) ||
+    stop("'log_path' must be a string: ", log_path)
 
   msg <- paste0(...)
 
@@ -120,16 +118,14 @@ warn <- function(
   msg_types = getOption("msgr.types"),
   log_path  = getOption("msgr.log_path"))
 {
-  {
-    (is_natural(level) && level <= 10) ||
-      stop("'level' must be an integer between 1 and 10: ", level)
-    (is_natural(msg_level) && msg_level <= 10) ||
-      stop("'msg_level' must be an integer between 1 and 10: ", msg_level)
-    ((is_null(msg_types) || is_character(msg_types)) && is_in(msg_types, c("INFO", "WARNING", "ERROR"))) ||
-      stop("'msg_types' must be NULL (no messages) or a character vector containing \"INFO\", \"WARNING\" or \"ERROR\": ", msg_types)
-    is_string(log_path) ||
-      stop("'log_path' must be a string: ", log_path)
-  }
+  (is_scalar_integerish(level) && isTRUE(level > 0) && isTRUE(level <= 10)) ||
+    stop("'level' must be an integer between 1 and 10: ", level)
+  (is_scalar_integerish(msg_level) && isTRUE(msg_level > 0) && isTRUE(msg_level <= 10)) ||
+    stop("'msg_level' must be an integer between 1 and 10: ", msg_level)
+  ((is_null(msg_types) || is_character(msg_types)) && is_in(msg_types, c("INFO", "WARNING", "ERROR"))) ||
+    stop("'msg_types' must be NULL (no messages) or a character vector containing \"INFO\", \"WARNING\" or \"ERROR\": ", msg_types)
+  (is_scalar_character(log_path)) ||
+    stop("'log_path' must be a string: ", log_path)
 
   msg <- paste0(...)
 
@@ -199,16 +195,14 @@ error <- function(
   msg_types = getOption("msgr.types"),
   log_path  = getOption("msgr.log_path"))
 {
-  {
-    (is_natural(level) && level <= 10) ||
-      stop("'level' must be an integer between 1 and 10: ", level)
-    (is_natural(msg_level) && msg_level <= 10) ||
-      stop("'msg_level' must be an integer between 1 and 10: ", msg_level)
-    ((is_null(msg_types) || is_character(msg_types)) && is_in(msg_types, c("INFO", "WARNING", "ERROR"))) ||
-      stop("'msg_types' must be NULL (no messages) or a character vector containing \"INFO\", \"WARNING\" or \"ERROR\": ", msg_types)
-    is_string(log_path) ||
-      stop("'log_path' must be a string: ", log_path)
-  }
+  (is_scalar_integerish(level) && isTRUE(level > 0) && isTRUE(level <= 10)) ||
+    stop("'level' must be an integer between 1 and 10: ", level)
+  (is_scalar_integerish(msg_level) && isTRUE(msg_level > 0) && isTRUE(msg_level <= 10)) ||
+    stop("'msg_level' must be an integer between 1 and 10: ", msg_level)
+  ((is_null(msg_types) || is_character(msg_types)) && is_in(msg_types, c("INFO", "WARNING", "ERROR"))) ||
+    stop("'msg_types' must be NULL (no messages) or a character vector containing \"INFO\", \"WARNING\" or \"ERROR\": ", msg_types)
+  (is_scalar_character(log_path)) ||
+    stop("'log_path' must be a string: ", log_path)
 
   msg <- paste0(...)
 
@@ -264,16 +258,14 @@ info_if <- function(
   msg_types = getOption("msgr.types"),
   log_path  = getOption("msgr.log_path"))
 {
-  {
-    (is_natural(level) && level <= 10) ||
-      error("'level' must be an integer between 1 and 10: ", level)
-    (is_natural(msg_level) && msg_level <= 10) ||
-      error("'msg_level' must be an integer between 1 and 10: ", msg_level)
-    ((is_null(msg_types) || is_character(msg_types)) && is_in(msg_types, c("INFO", "WARNING", "ERROR"))) ||
-      error("'msg_types' must be NULL (no messages) or a character vector containing \"INFO\", \"WARNING\" or \"ERROR\": ", msg_types)
-    is_string(log_path) ||
-      error("'log_path' must be a string: ", log_path)
-  }
+  (is_scalar_integerish(level) && isTRUE(level > 0) && isTRUE(level <= 10)) ||
+    error("'level' must be an integer between 1 and 10: ", level)
+  (is_scalar_integerish(msg_level) && isTRUE(msg_level > 0) && isTRUE(msg_level <= 10)) ||
+    error("'msg_level' must be an integer between 1 and 10: ", msg_level)
+  ((is_null(msg_types) || is_character(msg_types)) && is_in(msg_types, c("INFO", "WARNING", "ERROR"))) ||
+    error("'msg_types' must be NULL (no messages) or a character vector containing \"INFO\", \"WARNING\" or \"ERROR\": ", msg_types)
+  (is_scalar_character(log_path)) ||
+    error("'log_path' must be a string: ", log_path)
 
   if (isTRUE(condition)) {
     uneval_condition <- substitute(condition)
@@ -325,16 +317,14 @@ warn_if <- function(
   msg_types = getOption("msgr.types"),
   log_path  = getOption("msgr.log_path"))
 {
-  {
-    (is_natural(level) && level <= 10) ||
-      error("'level' must be an integer between 1 and 10: ", level)
-    (is_natural(msg_level) && msg_level <= 10) ||
-      error("'msg_level' must be an integer between 1 and 10: ", msg_level)
-    ((is_null(msg_types) || is_character(msg_types)) && is_in(msg_types, c("INFO", "WARNING", "ERROR"))) ||
-      error("'msg_types' must be NULL (no messages) or a character vector containing \"INFO\", \"WARNING\" or \"ERROR\": ", msg_types)
-    is_string(log_path) ||
-      error("'log_path' must be a string: ", log_path)
-  }
+  (is_scalar_integerish(level) && isTRUE(level > 0) && isTRUE(level <= 10)) ||
+    error("'level' must be an integer between 1 and 10: ", level)
+  (is_scalar_integerish(msg_level) && isTRUE(msg_level > 0) && isTRUE(msg_level <= 10)) ||
+    error("'msg_level' must be an integer between 1 and 10: ", msg_level)
+  ((is_null(msg_types) || is_character(msg_types)) && is_in(msg_types, c("INFO", "WARNING", "ERROR"))) ||
+    error("'msg_types' must be NULL (no messages) or a character vector containing \"INFO\", \"WARNING\" or \"ERROR\": ", msg_types)
+  (is_scalar_character(log_path)) ||
+    error("'log_path' must be a string: ", log_path)
 
   if (isTRUE(condition)) {
     uneval_condition <- substitute(condition)
@@ -386,16 +376,14 @@ error_if <- function(
   msg_types = getOption("msgr.types"),
   log_path  = getOption("msgr.log_path"))
 {
-  {
-    (is_natural(level) && level <= 10) ||
-      error("'level' must be an integer between 1 and 10: ", level)
-    (is_natural(msg_level) && msg_level <= 10) ||
-      error("'msg_level' must be an integer between 1 and 10: ", msg_level)
-    ((is_null(msg_types) || is_character(msg_types)) && is_in(msg_types, c("INFO", "WARNING", "ERROR"))) ||
-      error("'msg_types' must be NULL (no messages) or a character vector containing \"INFO\", \"WARNING\" or \"ERROR\": ", msg_types)
-    is_string(log_path) ||
-      error("'log_path' must be a string: ", log_path)
-  }
+  (is_scalar_integerish(level) && isTRUE(level > 0) && isTRUE(level <= 10)) ||
+    error("'level' must be an integer between 1 and 10: ", level)
+  (is_scalar_integerish(msg_level) && isTRUE(msg_level > 0) && isTRUE(msg_level <= 10)) ||
+    error("'msg_level' must be an integer between 1 and 10: ", msg_level)
+  ((is_null(msg_types) || is_character(msg_types)) && is_in(msg_types, c("INFO", "WARNING", "ERROR"))) ||
+    error("'msg_types' must be NULL (no messages) or a character vector containing \"INFO\", \"WARNING\" or \"ERROR\": ", msg_types)
+  (is_scalar_character(log_path)) ||
+    error("'log_path' must be a string: ", log_path)
 
   if (isTRUE(condition)) {
     uneval_condition <- substitute(condition)
@@ -447,16 +435,14 @@ assert <- function(
   msg_types = getOption("msgr.types"),
   log_path  = getOption("msgr.log_path"))
 {
-  {
-    (is_natural(level) && level <= 10) ||
-      error("'level' must be an integer between 1 and 10: ", level)
-    (is_natural(msg_level) && msg_level <= 10) ||
-      error("'msg_level' must be an integer between 1 and 10: ", msg_level)
-    ((is_null(msg_types) || is_character(msg_types)) && is_in(msg_types, c("INFO", "WARNING", "ERROR"))) ||
-      error("'msg_types' must be NULL (no messages) or a character vector containing \"INFO\", \"WARNING\" or \"ERROR\": ", msg_types)
-    is_string(log_path) ||
-      error("'log_path' must be a string: ", log_path)
-  }
+  (is_scalar_integerish(level) && isTRUE(level > 0) && isTRUE(level <= 10)) ||
+    error("'level' must be an integer between 1 and 10: ", level)
+  (is_scalar_integerish(msg_level) && isTRUE(msg_level > 0) && isTRUE(msg_level <= 10)) ||
+    error("'msg_level' must be an integer between 1 and 10: ", msg_level)
+  ((is_null(msg_types) || is_character(msg_types)) && is_in(msg_types, c("INFO", "WARNING", "ERROR"))) ||
+    error("'msg_types' must be NULL (no messages) or a character vector containing \"INFO\", \"WARNING\" or \"ERROR\": ", msg_types)
+  (is_scalar_character(log_path)) ||
+    error("'log_path' must be a string: ", log_path)
 
   if (!isTRUE(condition)) {
     uneval_condition <- substitute(condition)